% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{addLayersControlTitle}
\alias{addLayersControlTitle}
\title{Insert title above leafelt layer controls box}
\usage{
addLayersControlTitle(map, title)
}
\arguments{
\item{map}{leaflet htmlwidget object with layer control.}

\item{title}{character string to put above layer options.}
}
\value{
htmlwidget
}
\description{
Insert title above leafelt layer controls box
}
