% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{assign_heatmap_bins}
\alias{assign_heatmap_bins}
\title{Cut numeric column into colours bins for maps}
\usage{
assign_heatmap_bins(d, by, n_bins = 4, colours, sep = " - ")
}
\arguments{
\item{d}{data.frame or tibble.}

\item{by}{unquote numeric column name to split into bins.}

\item{n_bins}{the rough number of bins to create. Base R's pretty will give
or take to get 'pretty' break points.}

\item{colours}{character vector of colours for \code{colorRamp}, low values
of \code{by} will get assigned to the first values of colours while high
values get assigned to the tail of colours.}

\item{sep}{text separator between from and to values.}
}
\value{
A \code{tbl_df} with added columns fill_colour and label, plus list
  columns containing all_colours and all_labels used that can populate the
  map legend
}
\description{
\code{assign_heatmap_bins} takes a numeric column and breaks it into n equal
bins that are mapped to a colour palette and tidy labels.
}
