% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{chr_intervals}
\alias{chr_intervals}
\title{Convert breakpoints into nice from - to interval labels}
\usage{
chr_intervals(xs, sep)
}
\arguments{
\item{xs}{vector of break points}

\item{sep}{text separator between from and to values.}
}
\value{
A character vector of intervals
}
\description{
Convert breakpoints into nice from - to interval labels
}
