% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{escir_datatable}
\alias{escir_datatable}
\title{Custom datatable for GWRC reporting}
\usage{
escir_datatable(
  d,
  elementId = NULL,
  shared = FALSE,
  height = 360,
  container,
  cols_to_hide = c(),
  cols_to_left_align = c(),
  cols_to_fix_order = c(),
  cols_nonsortable = c(),
  col_widths = list(),
  col_colours = list(),
  escape = FALSE
)
}
\arguments{
\item{d}{data.frame or tibble.}

\item{elementId}{id of htmlwidget.}

\item{shared}{if TRUE, this will be linked via crosstalk to other widgets
(e.g. click and highlight), otherwise FALSE renders as simple interactive
table}

\item{height}{max height (px) before scrollbar used.}

\item{container}{manually specify html table layout for dataframe}

\item{cols_to_hide}{character vector of column names to suppress from output.}

\item{cols_to_left_align}{character vector of column names to left align}

\item{cols_to_fix_order}{parse character column to numeric for table sorting
(i.e. when a column has detection limits or states appended)}

\item{cols_nonsortable}{character vector of column names to disable sorting
buttons}

\item{col_widths}{named list or vector of column name = width, i.e.
list('Date' = '82px', 'Site' = '120px'). Use this if you need to stop a
column from text wrapping}

\item{col_colours}{named list or vector of column name = colour for
background shading, i.e. list('Good' = 'green', 'Bad' = 'red').}

\item{escape}{whether to escape HTML in the data (safe to leave as FALSE)}
}
\description{
\code{escir_datatable} is like the normal \code{DT::datatable} with some
defaults set and the options to hide/left align and not sort some columns.
}
\section{Notes}{


These columns are hidden by default:
\itemize{
  \item{key}
  \item{lng}
  \item{lat}
  \item{legend_colour}
  \item{fill_colour}
  \item{all_labels}
  \item{all_colours}
  \item{p_labs}
  \item{colour}
  \item{weight}
  \item{radius}
  \item{p_colours}
  \item{site_code}
  \item{map_label}
  \item{map_popup}
  \item{below_det}
  \item{survey_group}
}

These columns are left aligned by default:
\itemize{
  \item{Whaitua}
  \item{Site code}
  \item{Site}
  \item{Catchment}
  \item{Area}
  \item{Event}
  \item{Duration}
  \item{Site name}
  \item{Date}
  \item{Rainfall station}
  \item{Season}
  \item{Flows}
  \item{Land use}
  \item{Soil order}
  \item{Soil type}
  \item{Lake}
  \item{Type}
}
}

