% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazen_percentile.R
\name{hazen_percentile}
\alias{hazen_percentile}
\title{hazen_quantile ----}
\usage{
hazen_percentile(x, p = 0.95, na.rm = FALSE, strict = TRUE)
}
\arguments{
\item{x}{numeric vector with below detection limit values (x_i = '< lbd')
converted to half the detection limit, and above detection limit values
(x_i = '> ubd') converted to the detection limit plus one.}

\item{p}{double percent / 100}

\item{na.rm}{remove missing values}

\item{strict}{boolean, if true throw error on too few samples, otherwise yield NA}
}
\value{
Hazen percentile value
}
\description{
adapted from https://www.mfe.govt.nz/publications/fresh-water/bathewatch-user-guide/hazen-percentile-calculator
}
