% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hilltop.R
\name{htp_get_data}
\alias{htp_get_data}
\title{Get data for site/s - measurement/s for a date window}
\usage{
htp_get_data(
  url,
  Site,
  Measurement,
  From = NULL,
  To = NULL,
  verbose = FALSE,
  tstype = NULL,
  avg_by = NULL,
  just_i1 = FALSE,
  metadata = TRUE
)
}
\arguments{
\item{url}{Hilltop webservice url, this points to the desired .hts file.}

\item{Site}{site name/s to query.}

\item{Measurement}{measurements to query.}

\item{From}{date or datetime of first data point can be date, POSIXct, or
character string in the form YYYY-MM-DD or YYYY-MM-DD HH:MM:SS, see
examples.}

\item{To}{time of last measurement to get, see From format.}

\item{verbose}{boolean, whether to print the url query.}

\item{tstype}{if StdQualSeries required.}

\item{avg_by}{Hilltop averaging period, i.e. '1 hour', '1 day', will be
deprecated in future. Better to include this with the measurement name
i.e, Measurement = 'Flow&Method=Average&Interval=1 day$Align=1 day etc.
See https://www.hbrc.govt.nz/assets/Document-Library/Council-Data/20170426-HilltopServerTrimmed.pdf
for info on averaging methods like extrema and total.}

\item{just_i1}{faster parsing for continuous series}

\item{metadata}{whether to return metadata like units, series type etc.}
}
\value{
A \code{tbl_df} containing the data for the sites-measurements-times
  requested.
}
\description{
\code{htp_get_data} fetches data for the sites-measurements requested from
\code{From} to \code{To}.
}
\examples{
\dontrun{
# get hilltop web service url to desired data file
htp_url <- 'http://hilltop.{COUNCIL_ABBR}.govt.nz/{DATA_FILE}.hts?Service=Hilltop'

# discrete data
htp_get_data(htp_url,
             Site = 'SITEA',
             From = '2019-03-01',
             To = '2019-07-01',
             Measurement = 'Enterococci Bacteria')

# add just_i1 for continuous data for much faster results
# metadata false is already know units etc.
htp_get_data(htp_url,
             Site = 'SITEB',
             From = '2019-03-01',
             To = '2019-07-01',
             Measurement = 'Conductivity (TC)',
             just_i1 = TRUE,
             metadata = FALSE)
}
}
