% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hilltop.R
\name{htp_measurement_list}
\alias{htp_measurement_list}
\title{Get measurement information for sites}
\usage{
htp_measurement_list(url, Site = NULL, verbose = FALSE)
}
\arguments{
\item{url}{Hilltop webservice url, this points to the desired .hts file.}

\item{Site}{Hilltop site name/s to query.}

\item{verbose}{boolean, whether to print the url query.}
}
\value{
A \code{tbl_df} containing the sites and measurement metadata.
}
\description{
\code{htp_measurement_list} fetches available measurements for the sites
requested.
}
\examples{
\dontrun{
# get hilltop web service url to desired data file
htp_url <- 'http://hilltop.{COUNCIL_ABBR}.govt.nz/{DATA_FILE}.hts?Service=Hilltop'

htp_measurement_list(htp_url)
htp_measurement_list(htp_url, 'SITEA')
htp_measurement_list(htp_url, c('SITEA', 'SITEB'))
}
}
