% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw_grid.R
\name{hw_grid_2}
\alias{hw_grid_2}
\title{Lays out highchart widgets into a "grid", similar to
\code{grid.arrange} from \code{gridExtra}.}
\usage{
hw_grid_2(
  ...,
  ncol = NULL,
  rowheight = NULL,
  add_htmlgrid_css = TRUE,
  browsable = TRUE
)
}
\arguments{
\item{...}{either individual `highchart` objects or a mixture of
individual `highchart` objects and `list`s of `highchart` objects.}

\item{ncol}{how many columns in the grid}

\item{rowheight}{Height in px.}

\item{add_htmlgrid_css}{A logical value to add or not `htmlgrid.css` as
dependency.}

\item{browsable}{Logical value indicating if the returned object is converted
  to an HTML object browsable using \code{htmltools::browsable}.

This is a hack to fix the bootstrap issues with highcharter::hw_grid}
}
\description{
Lays out highchart widgets into a "grid", similar to
\code{grid.arrange} from \code{gridExtra}.
}
