% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pH_adjusted_NH3_N}
\alias{pH_adjusted_NH3_N}
\title{Use pH to adjust ammoniacal nitrogen}
\usage{
pH_adjusted_NH3_N(pH, NH3_N)
}
\arguments{
\item{pH}{vector or column of pH values}

\item{NH3_N}{vector or column of ammoniacal nitrogen values}
}
\value{
vector of NH4-N values
}
\description{
\code{pH_adjusted_NH3_N} estimates NH4-N, the nitrogen concentration of the
ammonium ion based on a pH of 8 and water temperature of 20degC. GWRC method.
}
\examples{
pH_adjusted_NH3_N(pH = rnorm(10, 7, 1), NH3_N = rep(0.01))
}
