% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_format}
\alias{safe_format}
\title{Format numbers with safe rounding}
\usage{
safe_format(x, digits)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of decimal places, if 0 this will add commas between thousands.}
}
\value{
rounded and formatted vector
}
\description{
\code{safe_format} uses base R format with safer rounding (see \code{?safe_round}). Useful for rounding a numeric column for outputs that want trailing zeroes kept for aligment. It doesn't use scientific representation on large numbers.
}
\examples{
\dontrun{
x <- c(0.049999, 0.05, 0.050001)
# base
format(x, digits = 1)
#> "0.05" "0.05" "0.05"
safe_format(x, digits = 1)
#> "0.0" "0.1" "0.1"

x <- c(1500.499, 1500.5, 1500.5001, 15000.5)
# base
format(x, digits = 0)
#> "1.500499e+03" "1.500500e+03" "1.500500e+03" "1.500050e+04"
# safe
safe_format(x, digits = 0)
#> "1,500" "1,500" "1,501" "15,001"
}
}
