% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_round}
\alias{safe_round}
\title{Round a value as you think it would be rounded}
\usage{
safe_round(x, digits)
}
\arguments{
\item{x}{numeric vector to be rounded.}

\item{digits}{number of decimal places to round to.}
}
\value{
rounded vector.
}
\description{
\code{safe_round} rounds numeric vectors to the digits specified. This is
recommended over base round as for rounding off a 5, the IEC 60559 standard
(see also ‘IEEE 754’) is expected to be used, ‘go to the even digit’.
Therefore round(0.5) is 0 and round(-1.5) is -2. However, this is dependent
on OS services and on representation error (since e.g. 0.15 is not
represented exactly, the rounding rule applies to the represented number and
not to the printed number, and so round(0.15, 1) could be either 0.1 or 0.2).
So add smallest possible number to x to ensure 0.0005 => 0.001 and similar.
}
\examples{
\dontrun{
x <- c(0.049999, 0.05, 0.050001)
# base rounds 0.05 down to 0
round(x, 1)
#> 0.0 0.0 0.1
# safe_round goes to 0.1
safe_round(x, 1)
#> 0.0 0.1 0.1
}
}
