% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hilltop.R
\name{htp_get_data}
\alias{htp_get_data}
\title{Get data for site/s - measurement/s for a date window}
\usage{
htp_get_data(
  url,
  Site,
  Measurement,
  From = NULL,
  To = NULL,
  verbose = FALSE,
  tstype = NULL,
  avg_by = NULL,
  suppress_warnings = TRUE,
  just_i1 = FALSE,
  metadata = TRUE,
  ...
)
}
\arguments{
\item{url}{Hilltop webservice url, this points to the desired .hts file.}

\item{Site}{site name/s to query.}

\item{Measurement}{measurements to query.}

\item{From}{date or datetime of first data point can be date, POSIXct, or
character string in the form YYYY-MM-DD or YYYY-MM-DD HH:MM:SS, see
examples.}

\item{To}{time of last measurement to get, see From format.}

\item{verbose}{boolean, whether to print the url query.}

\item{tstype}{if StdQualSeries required.}

\item{avg_by}{Hilltop averaging period, i.e. '1 hour', '1 day'}

\item{suppress_warnings}{TRUE (default) hides warnings for when there is no
data for the parameters requested.}

\item{just_i1}{faster parsing for continuous series}

\item{metadata}{whether to return metadata like units, series type etc.}

\item{...}{Currently unused.}
}
\value{
A \code{tbl_df} containing the data for the sites-measurements-times
  requested.
}
\description{
\code{htp_get_data} fetches data for the sites-measurements requested from
\code{From} to \code{To}.
}
\examples{
\dontrun{
htp_get_data('http://hilltop.gw.govt.nz/data.hts?Service=Hilltop',
             Site = 'Owhiro Bay',
             From = '2019-03-01',
             To = '2019-07-01',
             Measurement = 'Enterococci Bacteria')
#> [1] "Enterococci Bacteria - Owhiro Bay"
#> # A tibble: 4 x 22
#> site  timestamp           value lab sample num~ method gwrc programme
#> <chr> <dttm>              <chr> <chr>            <chr>  <chr>
#> Owhi~ 2019-03-04 11:20:00 12    19/10788-21      Enter~ Rec WQ
#> Owhi~ 2019-03-11 11:15:00 4     19/11079-21      Enter~ Rec WQ
#> Owhi~ 2019-03-18 10:34:00 4     19/12378-21      Enter~ Rec WQ
#> Owhi~ 2019-03-25 11:02:00 24    19/13658-21      Enter~ Rec WQ
#> # ... with 16 more variables
}
}
