% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hilltop.R
\name{htp_measurement_list}
\alias{htp_measurement_list}
\title{Get measurement information for sites}
\usage{
htp_measurement_list(url, Site = NULL, verbose = FALSE, ...)
}
\arguments{
\item{url}{Hilltop webservice url, this points to the desired .hts file.}

\item{Site}{Hilltop site name/s to query.}

\item{verbose}{boolean, whether to print the url query.}

\item{...}{Currently unused.}
}
\value{
A \code{tbl_df} containing the sites and measurement metadata.
}
\description{
\code{htp_measurement_list} fetches available measurements for the sites
requested.
}
\examples{
\dontrun{
htp_measurement_list('http://hilltop.gw.govt.nz/data.hts?Service=Hilltop',
                     Site = 'Abbots Creek at Featherston')
#> # A tibble: 1 x 7
#> inactive northing firstsynonym   recordingauthor~ comment     metnumber easting
#> <chr>    <chr>    <chr>          <chr>            <chr>       <chr>     <chr>
#> 0        5445581  Abbots Creek ~ GW-Eastern       "Site name~ D15131    1795280
}
}
