% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hilltop.R
\name{htp_site_info}
\alias{htp_site_info}
\title{Get information for a single site}
\usage{
htp_site_info(url, Site, verbose = FALSE, ...)
}
\arguments{
\item{url}{Hilltop webservice url, this points to the desired .hts file.}

\item{Site}{Hilltop site name to query.}

\item{verbose}{boolean, whether to print the url query.}

\item{...}{Currently unused.}
}
\value{
A \code{tbl_df} containing the site metadata.
}
\description{
\code{htp_site_info} fetches metadata for a single Hilltop site. See
\code{?htp_site_list} for multiple sites.
}
\examples{
\dontrun{
htp_site_info('http://hilltop.gw.govt.nz/data.hts?Service=Hilltop',
              Site = 'Abbots Creek at Featherston')
#> # A tibble: 1 x 7
#> inactive northing firstsynonym  recordingaut~ comment          metnumber easting
#> <chr>    <chr>    <chr>         <chr>         <chr>            <chr>     <chr>
#> 0        5445581  Abbots Creek~ GW-Eastern    "Site named \"F~ D15131    1795280
}
}
