% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hilltop.R
\name{htp_site_list}
\alias{htp_site_list}
\title{Get site lists}
\usage{
htp_site_list(url, Location = NULL, Measurement = NULL, verbose = FALSE)
}
\arguments{
\item{url}{Hilltop webservice url, this points to the desired .hts file.}

\item{Location}{'Yes' returns NZTM easting and northing. 'LatLong' latitude
and longitude. NULL returns no location columns.}

\item{Measurement}{Filter to only sites with this measurement.}

\item{verbose}{boolean, whether to print the url query.}
}
\value{
A \code{tbl_df} containing the site names and locations/measurements
  if requested
}
\description{
\code{htp_site_list} fetches site names and metadata from Hilltop
}
\examples{
\dontrun{
# get hilltop web service url to desired data file
htp_url <- 'http://hilltop.{COUNCIL_ABBR}.govt.nz/{DATA_FILE}.hts?Service=Hilltop'

htp_site_list(htp_url)
htp_site_list(htp_url, Location = 'Yes')
htp_site_list(htp_url, Location = 'LatLong')

htp_site_list(htp_url, Measurement = c('E-Coli'))
htp_site_list(htp_url, Location = 'Yes', Measurement = c('E-Coli'))
htp_site_list(htp_url, Location = 'LatLong', Measurement = c('E-Coli'))

htp_site_list(htp_url, Measurement = c('E-Coli', 'Rainfall'))
htp_site_list(htp_url, Location = 'Yes', Measurement = c('E-Coli', 'Rainfall'))
htp_site_list(htp_url, Location = 'LatLong', Measurement = c('E-Coli', 'Rainfall'))
}
}
